refactor(model): one place per question -- shared typed queries over the model
Five review rounds on the `causes:` engine produced the same defect
four times in different clothes: a rule settled in one place and a
sibling query left on the old key. Delivery joined on the wire
subject while the route query still joined on the syntactic
`declared_topic`; uncertainty was scoped to an endpoint in one check
and global in another. Each of those is ONE question the model can
answer once, so the questions move out of the engines:

  * `may_deliver` -- can this publish reach this subscription? Wire
    identity, wildcard coverage, and exact key domains decided;
    everything unknown widens.
  * `endpoint_incomplete` -- is what lies beyond this endpoint fully
    modeled? Scoped holes AND typed routes, in one answer.
  * `effects_of` -- what does this fn definitely do, and is that all?
    (Never `Function::effects`, which is rendered and lossy.)
  * `effect_class_atoms` / `render_effect_classes` -- what does a
    class name mean, and how is a set of them spelled?

`endpoint_incomplete` is DIRECTIONAL, which is what makes it shared
rather than merely moved. `causes:` asks what a publish can reach;
`depends:` asks what can reach a subscription. The completeness
question is the same shape on both sides and so are the ways it can
fail: downstream reads SUBSCRIBES|KEY_FILTERS|DELIVERY holes and a
`connect` route; upstream reads PUBLISHES|DELIVERY holes and a
`listen` route -- an inbound binding means an external peer may
publish into this subject, the exact mirror of the outbound finding
from round 4 that cost two review rounds to notice.

`judge_causes` now asks these instead of carrying its own copies;
its 23 controls pass unchanged, which is the point -- this moves no
behaviour. The re-exports keep `judgment::may_deliver` and friends
working for callers and tests.
